+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - [0007346](https://bugs.oxid-esales.com/view.php?id=7346) Update configuration to be able to force 3DSecure check for ACDC payments. Ensure 3D check result is validated depending on configuration.
 - PUI order in case of invalid phone number will now stay on order page and user can retry.
+- Fixed missing installemnt banners and shop start page, search etc.
 
 ## [2.1.6] - 2022-08-05